Enabled visualization.

+++ worldmodel.h
@@ -140,7 +140,7 @@ class cWorldModel : public cDriver::cHandler
     boost::asio::ip::tcp::socket socket;
     bool activated;
     double virtualStart;
-    double confidence;
+    float confidence;
   };
 
   class cOffsetSerializor : public cListener
